+++ tests/actor/test_actor.py
@@ -16,7 +16,7 @@ from dapr.actor.runtime.runtime import ActorRuntime
 from dapr.actor.runtime.typeinformation import ActorTypeInformation
 from dapr.serializers import DefaultJSONSerializer
 
-from tests.actor.fakeactorclasses import (
+from tests.actor.fake_actor_classes import (
     FakeSimpleActor,
     FakeSimpleReminderActor,
     FakeSimpleTimerActor,
