+++ tests/actor/test_state_manager.py
@@ -18,7 +18,7 @@ from dapr.actor.runtime.typeinformation import ActorTypeInformation
 from dapr.clients import DaprActorClientBase
 from dapr.serializers import DefaultJSONSerializer
 
-from tests.actor.fakeactorclasses import FakeSimpleActor
+from tests.actor.fake_actor_classes import FakeSimpleActor
 
 
 class ActorStateManagerTests(unittest.IsolatedAsyncioTestCase):
